+++ heroku/modules/translations.py
@@ -90,6 +90,7 @@ class Translations(loader.Module):
     @loader.command()
     async def setlang(self, message: Message):
         if not (args := utils.get_args_raw(message).lower()):
+
             await self._choose_language(message=message)
             return